--- tests/configs/standard/output/cm-opt-lsst-software-jupyterlab-panda.txt
@@ -1,17 +1,0 @@
-# Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
-#
-# Authors:
-# - Wen Guan, <wen.guan@cern.ch>, 2020
-[common]
-# if logdir is configured, idds will write to idds.log in this directory.
-# else idds will go to stdout/stderr.
-# With supervisord, it's good to write to stdout/stderr, then supervisord can manage and rotate logs.
-# logdir = /var/log/idds
-loglevel = INFO
-[rest]
-host = https://iddsserver.cern.ch:443/idds
-#url_prefix = /idds
-#cacher_dir = /tmp
-cacher_dir = /data/idds
