--- tests/configs/standard/output/nss-etc-group.txt
@@ -1,38 +1,0 @@
-root:x:0:
-bin:x:1:
-daemon:x:2:
-sys:x:3:
-adm:x:4:
-tty:x:5:
-disk:x:6:
-lp:x:7:
-mem:x:8:
-kmem:x:9:
-wheel:x:10:
-cdrom:x:11:
-mail:x:12:
-man:x:15:
-dialout:x:18:
-floppy:x:19:
-games:x:20:
-utmp:x:22:
-tape:x:33:
-utempter:x:35:
-video:x:39:
-ftp:x:50:
-lock:x:54:
-tss:x:59:
-audio:x:63:
-dbus:x:81:
-screen:x:84:
-nobody:x:99:
-users:x:100:
-systemd-journal:x:190:
-systemd-network:x:192:
-cgred:x:997:
-ssh_keys:x:998:
-input:x:999:
-rachel:x:1101:
-lunatics:x:2028:rachel
-mechanics:x:2001:rachel
-storytellers:x:2021:rachel
