--- tests/configs/standard/output/pod.json
@@ -1,253 +1,0 @@
-{
-  "containers": [
-    {
-      "args": [
-        "/opt/lsst/software/jupyterlab/runlab.sh"
-      ],
-      "env": [
-        {
-          "name": "K8S_NODE_NAME",
-          "value_from": {
-            "field_ref": {
-              "field_path": "spec.nodeName"
-            }
-          }
-        }
-      ],
-      "env_from": [
-        {
-          "config_map_ref": {
-            "name": "nb-rachel-env"
-          }
-        }
-      ],
-      "image": "lighthouse.ceres/library/sketchbook:latest_daily@sha256:1234",
-      "image_pull_policy": "Always",
-      "name": "notebook",
-      "ports": [
-        {
-          "container_port": 8888,
-          "name": "jupyterlab"
-        }
-      ],
-      "resources": {
-        "limits": {
-          "cpu": "1.0",
-          "memory": "3221225472"
-        },
-        "requests": {
-          "cpu": "0.25",
-          "memory": "805306368"
-        }
-      },
-      "security_context": {
-	"run_as_group": 1101,
-        "run_as_non_root": true,
-        "run_as_user": 1101
-      },
-      "volume_mounts": [
-        {
-          "mount_path": "/home",
-          "name": "home",
-          "read_only": false
-        },
-        {
-          "mount_path": "/project",
-          "name": "project",
-          "read_only": true
-        },
-        {
-          "mount_path": "/scratch",
-          "name": "scratch",
-          "read_only": false
-        },
-        {
-          "mount_path": "/etc/passwd",
-          "name": "nss-rachel-passwd",
-          "read_only": true,
-          "sub_path": "passwd"
-        },
-        {
-          "mount_path": "/etc/group",
-          "name": "nss-rachel-group",
-          "read_only": true,
-          "sub_path": "group"
-        },
-        {
-          "mount_path": "/opt/lsst/software/jupyterlab/lsst_dask.yml",
-          "name": "nss-rachel-lsst-dask-yml",
-          "read_only": true,
-          "sub_path": "lsst_dask.yml"
-        },
-        {
-          "mount_path": "/opt/lsst/software/jupyterlab/panda",
-          "name": "nss-rachel-panda",
-          "read_only": true,
-          "sub_path": "panda"
-        },
-        {
-          "mount_path": "/opt/lsst/software/jupyterlab/secrets",
-          "name": "nb-rachel-secrets",
-          "read_only": true
-        },
-        {
-          "mount_path": "/opt/lsst/software/jupyterlab/environment",
-          "name": "nb-rachel-env",
-          "read_only": false
-        },
-        {
-          "mount_path": "/tmp",
-          "name": "tmp",
-          "read_only": false
-        },
-        {
-          "mount_path": "/opt/lsst/software/jupyterlab/runtime",
-          "name": "nb-rachel-runtime",
-          "read_only": true
-        }
-      ],
-      "working_dir": "/home/rachel"
-    }
-  ],
-  "init_containers": [],
-  "restart_policy": "OnFailure",
-  "security_context": {
-    "fs_group": 1101,
-    "run_as_non_root": true,
-    "supplemental_groups": [
-      1101,
-      2028,
-      2001,
-      2021
-    ]
-  },
-  "volumes": [
-    {
-      "name": "home",
-      "nfs": {
-        "path": "/share1/home",
-        "read_only": false,
-        "server": "10.13.105.122"
-      }
-    },
-    {
-      "name": "project",
-      "nfs": {
-        "path": "/share1/project",
-        "read_only": true,
-        "server": "10.13.105.122"
-      }
-    },
-    {
-      "name": "scratch",
-      "nfs": {
-        "path": "/share1/scratch",
-        "read_only": false,
-        "server": "10.13.105.122"
-      }
-    },
-    {
-      "config_map": {
-        "items": [
-          {
-            "key": "_._etc_._passwd",
-            "mode": 420,
-            "path": "passwd"
-          }
-        ],
-        "name": "nb-rachel-nss"
-      },
-      "name": "nss-rachel-passwd"
-    },
-    {
-      "config_map": {
-        "items": [
-          {
-            "key": "_._etc_._group",
-            "mode": 420,
-            "path": "group"
-          }
-        ],
-        "name": "nb-rachel-nss"
-      },
-      "name": "nss-rachel-group"
-    },
-    {
-      "config_map": {
-        "items": [
-          {
-            "key": "_._opt_._lsst_._software_._jupyterlab_._lsst_dask.yml",
-            "mode": 420,
-            "path": "lsst_dask.yml"
-          }
-        ],
-        "name": "nb-rachel-configmap"
-      },
-      "name": "nss-rachel-lsst-dask-yml"
-    },
-    {
-      "config_map": {
-        "items": [
-          {
-            "key": "_._opt_._lsst_._software_._jupyterlab_._panda",
-            "mode": 420,
-            "path": "panda"
-          }
-        ],
-        "name": "nb-rachel-configmap"
-      },
-      "name": "nss-rachel-panda"
-    },
-    {
-      "name": "nb-rachel-secrets",
-      "secret": {
-        "secret_name": "nb-rachel"
-      }
-    },
-    {
-      "config_map": {
-        "name": "nb-rachel-env"
-      },
-      "name": "nb-rachel-env"
-    },
-    {
-      "empty_dir": {},
-      "name": "tmp"
-    },
-    {
-      "downward_api": {
-        "items": [
-          {
-            "path": "limits_cpu",
-            "resource_field_ref": {
-              "container_name": "notebook",
-              "resource": "limits.cpu"
-            }
-          },
-          {
-            "path": "requests_cpu",
-            "resource_field_ref": {
-              "container_name": "notebook",
-              "resource": "requests.cpu"
-            }
-          },
-          {
-            "path": "limits_memory",
-            "resource_field_ref": {
-              "container_name": "notebook",
-              "resource": "limits.memory"
-            }
-          },
-          {
-            "path": "requests_memory",
-            "resource_field_ref": {
-              "container_name": "notebook",
-              "resource": "requests.memory"
-            }
-          }
-        ]
-      },
-      "name": "nb-rachel-runtime"
-    }
-  ]
-}
