+++ tests/handlers/labs_test.py
@@ -2,13 +2,59 @@
 
 from __future__ import annotations
 
+import json
+from pathlib import Path
+from typing import Any
+
 import pytest
 from httpx import AsyncClient
 
+from jupyterlabcontroller.config import Configuration
 from jupyterlabcontroller.constants import DROPDOWN_SENTINEL_VALUE
 from jupyterlabcontroller.factory import Factory
 
 from ..settings import TestObjectFactory
+from ..support.kubernetes import MockLabKubernetesApi
+
+
+def strip_none(model: dict[str, Any]) -> dict[str, Any]:
+    """Strip `None` values from a serialized Kubernetes object.
+
+    Comparing Kubernetes objects against serialized expected output is a bit
+    of a pain, since Kubernetes objects often contain tons of optional
+    parameters and the ``to_dict`` serialization includes every parameter.
+    The naive result is therefore tedious to read or understand.
+
+    This function works around this by taking a serialized Kubernetes object
+    and dropping all of the parameters that are set to `None`. The ``to_dict``
+    form of a Kubernetes object should be passed through it first before
+    comparing to the expected output.
+
+    Parmaters
+    ---------
+    model
+        Kubernetes model serialized with ``to_dict``.
+
+    Returns
+    -------
+    dict
+        Cleaned-up model with `None` parameters removed.
+    """
+    result = {}
+    for key, value in model.items():
+        if value is None:
+            continue
+        if isinstance(value, dict):
+            value = strip_none(value)
+        elif isinstance(value, list):
+            list_result = []
+            for item in value:
+                if isinstance(item, dict):
+                    item = strip_none(item)
+                list_result.append(item)
+            value = list_result
+        result[key] = value
+    return result
 
 
 @pytest.mark.asyncio
@@ -104,3 +150,31 @@ async def test_lab_start_stop(
         headers={"X-Auth-Request-User": user.username},
     )
     assert r.status_code == 404
+
+
+@pytest.mark.asyncio
+async def test_lab_objects(
+    client: AsyncClient,
+    config: Configuration,
+    mock_kubernetes: MockLabKubernetesApi,
+    obj_factory: TestObjectFactory,
+    std_result_dir: Path,
+) -> None:
+    token, user = obj_factory.get_user()
+    lab = obj_factory.labspecs[0]
+
+    r = await client.post(
+        f"/nublado/spawner/v1/labs/{user.username}/create",
+        json={"options": lab.options.dict(), "env": lab.env},
+        headers={
+            "X-Auth-Request-Token": token,
+            "X-Auth-Request-User": user.username,
+        },
+    )
+    assert r.status_code == 201
+
+    namespace = f"{config.lab.namespace_prefix}-{user.username}"
+    objects = mock_kubernetes.get_all_objects_in_namespace_for_test(namespace)
+    with (std_result_dir / "lab-objects.json").open("r") as f:
+        expected = json.load(f)
+    assert [strip_none(o.to_dict()) for o in objects] == expected
